# docs/swagger/swagger.json
{
  "basePath": "/",
  "definitions": {
    "V1BearerAuthError": {
      "properties": {
        "error": {
          "properties": {
            "message": {
              "example": "unauthenticated",
              "type": "string"
            }
          },
          "required": [
            "message"
          ],
          "type": "object"
        }
      },
      "required": [
        "error"
      ],
      "type": "object"
    },
    "V1CapabilityNotConfiguredError": {
      "properties": {
        "error": {
          "properties": {
            "capability": {
              "type": "string"
            },
            "code": {
              "example": "capability_not_configured",
              "type": "string"
            },
            "implemented": {
              "example": false,
              "type": "boolean"
            },
            "message": {
              "type": "string"
            }
          },
          "required": [
            "code",
            "message",
            "capability",
            "implemented"
          ],
          "type": "object"
        }
      },
      "required": [
        "error"
      ],
      "type": "object"
    },
...
    "V1OperatorListEnvelope": {
      "properties": {
        "items": {
          "items": {},
          "type": "array"
        },
        "meta": {
          "properties": {
            "limit": {
              "format": "int32",
              "type": "integer"
            },
            "returned": {
              "type": "integer"
            }
          },
          "required": [
            "limit",
            "returned"
          ],
          "type": "object"
        }
      },
      "required": [
        "items",
        "meta"
      ],
      "type": "object"
    },
...
    "V1StandardError": {
      "properties": {
        "error": {
          "properties": {
            "code": {
              "example": "invalid_json",
              "type": "string"
            },
            "message": {
              "type": "string"
            }
          },
          "required": [
            "code",
            "message"
          ],
          "type": "object"
        }
      },
      "required": [
        "error"
      ],
      "type": "object"
    },
...
  },
  "host": "localhost:8080",
  "info": {
    "contact": {
      "name": "AVF Engineering",
      "url": "https://github.com/avf/avf-vending-system"
    },
    "description": "HTTP API for the AVF vending platform (cmd/api). All `/v1/*` routes require `Authorization: Bearer <JWT>` unless noted (health, optional metrics, Swagger when enabled). Bearer middleware failures use minimal JSON `{\"error\":{\"message\":\"...\"}}` without `error.code`. Most `/v1` handlers use `{\"error\":{\"code\":\"...\",\"message\":\"...\"}}` (see writeAPIError). **501** list stubs return `not_implemented` plus `capability` and `implemented:false`. **503** `capability_not_configured` is used when optional wiring (MQTT dispatch, commerce persistence, outbox defaults) is missing. Request tracing: responses echo `X-Request-ID` and `X-Correlation-ID` when middleware is enabled.",
    "license": {
      "name": "License terms apply per your deployment agreement; see repository NOTICE if present."
    },
    "termsOfService": "https://github.com/avf/avf-vending-system/tree/main/avf-vending-api",
    "title": "AVF Vending HTTP API",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/machines/{machineId}/commands/dispatch": {
      "post": {
        "consumes": [
          "json"
        ],
        "description": "Requires **Idempotency-Key** or **X-Idempotency-Key**. Org/platform admin only. When MQTT publisher is missing \u2192 **503** `capability_not_configured` (`mqtt_command_dispatch`). Nil RemoteCommands wiring \u2192 **500**. Rate limiter may return **429** `rate_limited`.",
        "parameters": [
          {
            "description": "Machine UUID",
            "in": "path",
            "name": "machineId",
            "required": true,
            "type": "string"
          },
          {
            "description": "command_type, payload, desired_state, optional correlation_id, operator_session_id",
            "in": "body",
            "name": "body",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "Dispatch remote MQTT command",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "429": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "503": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1CapabilityNotConfiguredError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Dispatch remote MQTT command",
        "tags": [
          "Device"
        ]
      }
    },
    "/v1/machines/{machineId}/commands/receipts": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Mounted when RemoteCommands service is configured. If app wiring is nil, handler returns **500** `internal` (not 503). Optional **limit** query (default 50, max 500).",
        "parameters": [
          {
            "description": "Machine UUID",
            "in": "path",
            "name": "machineId",
            "required": true,
            "type": "string"
          },
          {
            "description": "Default 50, max 500",
            "in": "query",
            "name": "limit",
            "required": false,
            "type": "integer"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "List recent command receipts",
            "schema": {
              "$ref": "#/definitions/V1OperatorListEnvelope"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List recent command receipts",
        "tags": [
          "Device"
        ]
      }
    },
    "/v1/machines/{machineId}/commands/{sequence}/status": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "**sequence** must be a non-negative integer. Unknown sequence \u2192 **404** `command_not_found`.",
        "parameters": [
          {
            "description": "Machine UUID",
            "in": "path",
            "name": "machineId",
            "required": true,
            "type": "string"
          },
          {
            "description": "Ledger sequence (non-negative integer)",
            "in": "path",
            "name": "sequence",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "Get command dispatch status by sequence",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "404": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get command dispatch status by sequence",
        "tags": [
          "Device"
        ]
      }
    },
...
  },
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Send: Bearer <JWT>. JWT mode is selected by HTTP_AUTH_MODE (hs256 default, rs256_pem, rs256_jwks). 401/403 responses from this layer use JSON {\"error\":{\"message\":\"...\"}} without error.code.",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0",
...
}